examples/jailbreakbench-defense/requirements.txt
litellm>=1.50.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
